+++ Sources/Get/APIClient.swift
@@ -87,7 +87,7 @@ public actor APIClient {
     private func send<T>(_ request: Request<T>, _ decode: @escaping (Data) async throws -> T) async throws -> Response<T> {
         let response = try await data(for: request)
         let value = try await decode(response.value)
-        return response.map { _ in value }
+        return response.map { _ in value } // Keep metadata
     }
     
     /// Returns response data for the given request.
@@ -108,10 +108,10 @@ public actor APIClient {
     private func actuallySend(_ request: URLRequest) async throws -> Response<Data> {
         var request = request
         await delegate.client(self, willSendRequest: &request)
-        let (data, response) = try await loader.data(for: request, session: session)
+        let (data, response, metrics) = try await loader.data(for: request, session: session)
         try validate(response: response, data: data)
         let httpResponse = (response as? HTTPURLResponse) ?? HTTPURLResponse() // The right side should never be executed
-        return Response(value: data, data: data, request: request, response: httpResponse, statusCode: httpResponse.statusCode)
+        return Response(value: data, data: data, request: request, response: httpResponse, statusCode: httpResponse.statusCode, metrics: metrics)
     }
 
     private func makeRequest<T>(for request: Request<T>) async throws -> URLRequest {